pyproject.toml
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "fiboaitech"
version = "0.1.0"
description = "FiboAITech is an orchestration framework for agentic AI and LLM applications"
authors = ["FiboAITech Team <hello@getfiboaitech.ai>"]
license = "Apache License 2.0"
readme = "README.md"
keywords = ["ai", "gpt", "agents", "rag", "llm", "generative-ai", "llmops"]
homepage = "https://www.getfiboaitech.ai"
repository = "https://github.com/fiboaitech-ai/fiboaitech"

[tool.poetry.dependencies]
python = "~3.12"
jinja2 = "~3.1.4"
pydantic = "~2.7.1"
openai = "~1.45.0"
boto3 = "~1.34.34"
redis = "~5.0.0"
google-generativeai = "~0.5.0"
google-cloud-aiplatform = "~1.47.0"
litellm = "~1.46.0"
requests = "~2.31.0"
RestrictedPython = "~7.1"
jsonpath-ng = "~1.6.1"
jsonpickle = "~3.0.3"
pinecone-client = "~3.2.2"
chromadb-client = "~0.5.5.dev0"
unstructured-client = "~0.18.0"
omegaconf = "~2.3.0"
weaviate-client = "~4.7.1"
filetype = "~1.2.0"
more-itertools = "~10.3.0"
pdf2image = "~1.17.0"
e2b = "~0.17.1"
e2b-code-interpreter = "~0.0.10"
black = "~24.8.0"
pypdf = "~4.3.1"
python-pptx = "1.0.2"
qdrant-client = "~1.11.3"


[tool.poetry.group.dev.dependencies]
setuptools = "~69.1.1"
pre-commit = "~3.7.0"
pytest = "~8.1.1"
pytest-mock = "~3.14.0"
coverage = "~7.3.2"
fakeredis = "~2.21.3"
python-dotenv = "~1.0.1"
mkdocs = "~1.6.0"
mkdocs-material = "~9.5.27"
mkdocstrings = "~0.25.1"
mkdocstrings-python = "~1.10.4"
requests-mock = "~1.12.1"

[tool.poetry.group.examples]
optional = true

[tool.poetry.group.examples.dependencies]
fastapi = "~0.110.1"
uvicorn = "~0.25.0"
websockets = "~12.0"
sse-starlette = "~2.1.0"
httpx = "~0.27.0"
httpx-sse = "~0.4.0"
chainlit = "~1.2.0"
typer = "~0.12.3"
streamlit = "~1.36.0"
streamlit-date-picker = "~0.0.3"
streamlit-image-select = "~0.6.0"
streamlit-pdf-viewer = "~0.0.14"
weasyprint = "~62.3"
markdown = "~3.6"
beautifulsoup4 = "~4.12.3"
selenium = "~4.21.0"
tqdm = "~4.66.4"
sympy = "~1.13.1"
flagembedding = "~1.2.11"
pygraphviz = "~1.13"
agentops = "~0.3.12"
langfuse = "^2.51.2"
peft = "^0.13.0"
